Moved exceptions to own module.
